Use m0's deep sleep mode
This commit adds support for the m0's RTC. We can use the RTC to
schedule wakeups. We also use the interrupt from the radio to wake up
the controller.

Also included in this commit are options in config.h for selecting the
behavior of the LED. The LED can be set to turn on when the controller
is woken up, or when the radio is transmitting.